requirements.txt
# Web Scraping
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2
scrapy==2.11.0
playwright==1.40.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Email Validation
email-validator==2.1.0
validate-email==1.3

# Google Services
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0

# Visualization
folium==0.15.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Scheduling and Automation
schedule==1.2.0
apscheduler==3.10.4

# Anti-Bot Protection
fake-useragent==1.4.0
requests-html==0.10.0
undetected-chromedriver==3.5.4

# Configuration and Logging
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# Database (optional)
sqlite3

# Proxy Support
requests[socks]==2.31.0
pysocks==1.7.1

# Geocoding
geopy==2.4.0

# Data Validation
cerberus==1.3.5 